+++ src/modules/layerInformation/store/actionsLayerInformation.js
@@ -104,7 +104,7 @@ export default {
         let metadata;
 
         commit("setDownloadLinks", null);
-        if (metaInfo.cswUrl && typeof metaInfo.metaId !== "undefined") {
+        if (metaInfo.cswUrl && typeof metaInfo.metaId !== "undefined" && metaInfo.metaId !== null) {
             try {
                 metadata = await getCswRecordById.getRecordById(metaInfo.cswUrl, metaInfo.metaId);
             }